requirements.txt
google-generativeai>=0.3.0
python-dotenv>=1.0.0
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0
